feat 지그재그 크롤링 후 DB저장, 카테고리, detail, histroy 저장까지 완료:
