+++ models/shop_type.py
@@ -1,0 +1,6 @@
+from enum import Enum
+class ShopType(Enum):
+    MUSINSA = "MUSINSA"
+    ZIGZAG = "ZIGZAG"
+    ABLY = "ABLY"
+    BRANDY = "BRANDY"
